feat: 검색 결과 클릭 시 해당 상세화면으로 이동하도록 구현
#S15P11D207-225

+++ frontend/modera/feature/home/src/main/kotlin/com/ssafy/modera/feature/home/HomeScreen.kt
@@ -37,7 +37,7 @@ fun HomeRoute(
     onCalendarClick: () -> Unit,
     onSettingsClick: () -> Unit,
     onCategoryClick: (Category) -> Unit,
-    onSearchResultClick: (Int) -> Unit,
+    onSearchResultClick: (Long) -> Unit,
     modifier: Modifier = Modifier,
     viewModel: HomeViewModel = hiltViewModel(),
 ) {
@@ -87,7 +87,7 @@ private fun HomeSuccessScreen(
     onRecentSearchClick: (String) -> Unit,
     onRecentSearchDelete: (String) -> Unit,
     onSearchDeactivate: () -> Unit,
-    onSearchResultClick: (Int) -> Unit,
+    onSearchResultClick: (Long) -> Unit,
     modifier: Modifier = Modifier,
 ) {
     val searchFocusRequester = remember { FocusRequester() }
